requirements.txt
anthropic>=0.42.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-cov>=6.0.0
